add button to nft alert
